moved prettier & husky to folder of themself
