+++ src/logic/husky/index.ts
@@ -1,0 +1,3 @@
+import Husky from './husky';
+
+export default Husky;
