+++ src/logic/prettier/index.ts
@@ -1,0 +1,3 @@
+import Prettier from './prettier';
+
+export default Prettier;
